[update]admin選択肢の削除

+++ app/controllers/admin/choices_controller.rb
@@ -1,5 +1,5 @@
 class Admin::ChoicesController < ApplicationController
-  before_action :check_answer_existence, only: [:edit, :update, :destroy]
+  # before_action :check_answer_existence, only: [:edit, :update, :destroy]
 
   def new
     @choice = Choice.new
@@ -22,6 +22,10 @@ class Admin::ChoicesController < ApplicationController
 
   def edit
     @choice = Choice.find(params[:id])
+    if @choice.answers.present?
+      flash[:alert] = "回答があるため、編集できません。"
+      redirect_to admin_choices_path
+    end
   end
 
   def update
@@ -37,12 +41,18 @@ class Admin::ChoicesController < ApplicationController
 
   def destroy
     @choice = Choice.find(params[:id])
-    if @choice.destroy!
-      flash[:notice] = "成功"
-      redirect_to admin_choices_path
+    if @choice.answers.empty?
+      if @choice.destroy!
+        flash[:notice] = "成功"
+        redirect_to admin_choices_path
+      else
+        flash.now[:alert] = "失敗"
+        render admin_choices_path
+      end
     else
-      flash.now[:alert] = "失敗"
-      render admin_choices_path
+      # 選択肢に回答がある場合は削除しないように
+      flash[:alert] = "この選択肢はすでに回答されているため、削除できません。"
+      redirect_to admin_choices_path
     end
   end
 
@@ -51,13 +61,13 @@ class Admin::ChoicesController < ApplicationController
       params.require(:choice).permit(:choice, :question_id)
     end
 
-    def check_answer_existence
-      @choice = Choice.find(params[:id])
-      if @Choice.answers.exists?
-        flash[:alert] = "既に回答が存在するため、編集や削除はできません。"
-        redirect_to admin_choices_path
-      end
-    end
+    # def check_answer_existence
+    #   @choice = Choice.find(params[:id])
+    #   if @Choice.answers.exists?
+    #     flash[:alert] = "既に回答が存在するため、編集や削除はできません。"
+    #     redirect_to admin_choices_path
+    #   end
+    # end
